// Arashyn-FE/src/lib/api/generated/models/grammar-detail-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Arashyn API
 * REST API for Arashyn Platform
 *
 * The version of the OpenAPI document: v1.0
 * Contact: touman876@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { GrammarFilterResponse } from './grammar-filter-response';
// May contain unused imports in some cases
// @ts-ignore
import type { GrammarNoteResponse } from './grammar-note-response';
// May contain unused imports in some cases
// @ts-ignore
import type { Group } from './group';

export interface GrammarDetailResponse {
    'id'?: string;
    'title'?: string;
    'language'?: GrammarDetailResponseLanguageEnum;
    'isPublic'?: boolean;
    'ownerId'?: string;
    'ownerName'?: string;
    'groups'?: Array<Group>;
    'notes'?: Array<GrammarNoteResponse>;
    'filters'?: Array<GrammarFilterResponse>;
}

export enum GrammarDetailResponseLanguageEnum {
    Vi = 'VI',
    En = 'EN',
    Ja = 'JA',
    Ko = 'KO',
    Zh = 'ZH'
}